PL-32: Add very basic behat tests.
These can be run from the base of the repo from inside the vm: `vendor/bin/behat --tags=javascript`

Behat is configured to test the D8 site by default.

The wait script was necessary since sometimes the react app wasn't loaded before the system would try to find the text.

The D7 content is random, but `ex` seems to frequently be a string we can search for.